Перенесла routes в __init__.py, удалила routes.py, удалила deja_vu_words.py, переименовала папку app в webapp
