--- app/config.py
@@ -1,5 +1,0 @@
-import os
-
-
-class Config(object):
-    SECRET_KEY = os.environ.get('SECRET_KEY') or 'you-will-never-guess'
