--- app/routes.py
@@ -1,17 +1,0 @@
-# Для сохранения всех путей url
-
-from flask import render_template
-from app import app
-from app.forms import LoginForm
-
-
-@app.route('/')
-@app.route('/index')
-def index():
-    return render_template('index.html', title='Home')
-
-
-@app.route('/login')
-def login():
-    form = LoginForm()
-    return render_template('login.html', title='Sign In', form=form)
