--- deja_vu_words.py
@@ -1,3 +1,0 @@
-# Запуск приложения через этот файл
-
-from app import app
