+++ webapp/forms.py
@@ -1,4 +1,4 @@
-# Классы веб-форм
+# В этом файле хранятся классы веб-форм.
 
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, BooleanField, SubmitField
